@@ -12,8 +12,8 @@
     "@testing-library/user-event": "^12.8.3",
     "dompurify": "^2.2.8",
     "marked": "^2.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^0.2.4"
   },
@@ -46,6 +46,6 @@
     ]
   },
   "devDependencies": {
-    "node-sass": "^5.0.0"
+    "node-sass": "^7.0.3"
   }
 }
